+++ crates/cli/src/commands/debug/mod.rs
@@ -1,3 +1,5 @@
 mod config;
+mod env;
 
 pub use config::*;
+pub use env::*;